# src/controllers/v3/openapi.yaml
components:
...
    schemas:
        OrganizationRefID:
            type: string
        OrganizationUnitRefID:
            type: string
        Dataset:
            properties:
                extForeignKey:
                    type: string
                name:
                    type: string
                license_title:
                    type: string
                security_class:
                    type: string
                view_audience:
                    type: string
                download_audience:
                    type: string
                record_publish_date:
                    type: string
                notes:
                    type: string
                title:
                    type: string
                isInCatalog:
                    type: string
                isDraft:
                    type: string
                contacts:
                    type: string
                extSource:
                    type: string
                extRecordHash:
                    type: string
                tags:
                    items:
                        type: string
                    type: array
                resources: {}
                organization:
                    $ref: '#/components/schemas/OrganizationRefID'
                organizationUnit:
                    $ref: '#/components/schemas/OrganizationUnitRefID'
            type: object
            additionalProperties: false
        BatchResult:
            properties:
                status:
                    type: number
                    format: double
                result:
                    type: string
                reason:
                    type: string
                id:
                    type: string
                ownedBy:
                    type: string
                childResults:
                    items:
                        $ref: '#/components/schemas/BatchResult'
                    type: array
            required:
                - status
                - result
            type: object
            additionalProperties: false
        DraftDataset:
            properties:
                name:
                    type: string
                license_title:
                    type: string
                security_class:
                    type: string
                    enum:
                        - HIGH-CABINET
                        - HIGH-CONFIDENTIAL
                        - HIGH-SENSITIVITY
                        - MEDIUM-SENSITIVITY
                        - MEDIUM-PERSONAL
                        - LOW-SENSITIVITY
                        - LOW-PUBLIC
                        - PUBLIC
                        - 'PROTECTED A'
                        - 'PROTECTED B'
                        - 'PROTECTED C'
                view_audience:
                    type: string
                    enum:
                        - Public
                        - Government
                        - 'Named users'
                        - 'Government and Business BCeID'
                download_audience:
                    type: string
                    enum:
                        - Public
                        - Government
                        - 'Named users'
                        - 'Government and Business BCeID'
                record_publish_date:
                    type: string
                notes:
                    type: string
                title:
                    type: string
                isInCatalog:
                    type: boolean
                isDraft:
                    type: boolean
                contacts:
                    type: string
                resources:
                    type: string
                tags:
                    items:
                        type: string
                    type: array
                organization:
                    $ref: '#/components/schemas/OrganizationRefID'
                organizationUnit:
                    $ref: '#/components/schemas/OrganizationUnitRefID'
            type: object
            additionalProperties: false
            example:
                name: my_sample_dataset
                license_title: 'Open Government Licence - British Columbia'
                security_class: PUBLIC
                view_audience: Public
                download_audience: Public
                record_publish_date: '2017-09-05'
                notes: 'Some notes'
                title: 'A title about my dataset'
                tags:
                    - tag1
                    - tag2
                organization: ministry-of-citizens-services
                organizationUnit: databc
        Gateway:
            properties:
                gatewayId:
                    type: string
                displayName:
                    type: string
            type: object
            additionalProperties: false
        Maybe_Scalars-at-String_:
            type: string
            nullable: true
        NamespaceInput:
            properties:
                displayName:
                    $ref: '#/components/schemas/Maybe_Scalars-at-String_'
                name:
                    $ref: '#/components/schemas/Maybe_Scalars-at-String_'
            type: object
        ActivityDetail:
            properties:
                id:
                    type: string
                message:
                    type: string
                params:
                    properties: {}
                    additionalProperties:
                        type: string
                    type: object
                activityAt: {}
                blob: {}
            required:
                - message
                - params
                - activityAt
            type: object
            additionalProperties: false
...
    securitySchemes:
        jwt:
            type: oauth2
            description: 'Authz Client Credential'
            flows:
                clientCredentials:
                    tokenUrl: 'https://token_endpoint'
                    scopes: {}
        portal:
            type: http
            description: 'Authz Portal Login'
            scheme: bearer
            bearerFormat: JWT
        openid:
            type: openIdConnect
            description: 'OIDC Login'
            openIdConnectUrl: 'https://well_known_endpoint'
info:
    title: 'APS Directory API'
    version: 3.0.0
    description: 'API Services Portal by BC Gov API Programme Services'
    license:
        name: MIT
    contact:
        name: 'BC Gov APS'
openapi: 3.0.0
paths:
...
    '/organizations/{org}/datasets/{name}':
        delete:
            operationId: delete-dataset
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BatchResult'
            description: "Delete a Dataset\n> `Required Scope:` Dataset.Manage"
            summary: 'Delete a dataset'
            tags:
                - 'API Directory'
            security:
                -
                    jwt:
                        - Dataset.Manage
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: name
                    required: true
                    schema:
                        type: string
        get:
            operationId: get-organization-dataset
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Dataset'
            description: "Get metadata about a Dataset that are available by API for this organization\n> `Required Scope:` Dataset.Manage"
            summary: 'Get Organization Dataset'
            tags:
                - 'API Directory'
            security:
                -
                    jwt:
                        - Dataset.Manage
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: name
                    required: true
                    schema:
                        type: string
...
    '/gateways/{gatewayId}/datasets':
        put:
            operationId: put-dataset
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/BatchResult'
            description: "Update metadata about a Dataset\n> `Required Scope:` Namespace.Manage"
            summary: 'Update Dataset'
            tags:
                - 'API Directory'
            security:
                -
                    jwt:
                        - Namespace.Manage
            parameters:
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/DraftDataset'
    '/gateways/{gatewayId}/datasets/{name}':
        get:
            operationId: get-dataset
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Dataset'
            description: "Get metadata about a Dataset\n> `Required Scope:` Namespace.Manage"
            summary: 'Get Dataset'
            tags:
                - 'API Directory'
            security:
                -
                    jwt:
                        - Namespace.Manage
            parameters:
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: name
                    required: true
                    schema:
                        type: string
    /gateways/report:
        get:
            operationId: report
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema: {}
            tags:
                - Gateways
            security:
                -
                    jwt: []
            parameters:
                -
                    in: query
                    name: ids
                    required: false
                    schema:
                        default: '[]'
                        type: string
    /gateways:
        get:
            operationId: gateway-list
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                items:
                                    type: string
                                type: array
            summary: 'List of Gateway IDs'
            tags:
                - Gateways
            security:
                -
                    jwt: []
            parameters: []
        post:
            operationId: create-gateway
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Gateway'
            description: 'Create a gateway'
            summary: 'Create Gateway'
            tags:
                - Gateways
            security:
                -
                    jwt: []
            parameters: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/NamespaceInput'
    '/gateways/{gatewayId}':
        get:
            operationId: namespace-profile
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Gateway'
            description: "Get details about the gateway, such as permissions for what the gateway is setup with.\n> `Required Scope:` Namespace.Manage"
            summary: 'Gateway Summary'
            tags:
                - Gateways
            security:
                -
                    jwt:
                        - Namespace.Manage
            parameters:
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
        delete:
            operationId: delete-namespace
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Gateway'
            description: "Delete the gateway\n> `Required Scope:` Namespace.Manage"
            summary: 'Delete Gateway'
            tags:
                - Gateways
            security:
                -
                    jwt:
                        - Namespace.Manage
            parameters:
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
                -
                    in: query
                    name: force
                    required: false
                    schema:
                        default: false
                        type: boolean
    '/gateways/{gatewayId}/activity':
        get:
            operationId: gateway-admin-activity
            responses:
                '200':
                    description: 'Activity[]'
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/ActivityDetail'
                                type: array
            description: '> `Required Scope:` Namespace.View'
            summary: 'Get administration activity for this gateway'
            tags:
                - Gateways
            security:
                -
                    jwt:
                        - Namespace.View
            parameters:
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
                -
                    in: query
                    name: first
                    required: false
                    schema:
                        default: 20
                        format: double
                        type: number
                -
                    in: query
                    name: skip
                    required: false
                    schema:
                        default: 0
                        format: double
                        type: number
...
    '/organizations/{org}/{orgUnit}/gateways/{gatewayId}':
        put:
            operationId: assign-namespace-to-organization
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    result: {type: string}
                                required:
                                    - result
                                type: object
            description: '> `Required Scope:` Namespace.Assign'
            tags:
                - Organizations
            security:
                -
                    jwt:
                        - Namespace.Assign
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: orgUnit
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
                -
                    in: query
                    name: enable
                    required: false
                    schema:
                        default: true
                        type: boolean
        delete:
            operationId: unassign-namespace-from-organization
            responses:
                '200':
                    description: Ok
                    content:
                        application/json:
                            schema:
                                properties:
                                    result: {type: string}
                                required:
                                    - result
                                type: object
            description: '> `Required Scope:` Namespace.Assign'
            tags:
                - Organizations
            security:
                -
                    jwt:
                        - Namespace.Assign
            parameters:
                -
                    in: path
                    name: org
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: orgUnit
                    required: true
                    schema:
                        type: string
                -
                    in: path
                    name: gatewayId
                    required: true
                    schema:
                        type: string
...
servers:
    -
        url: /ds/api/v3
...